Move unblock, unescape and uncurly to Meta
